fix(stack-analyses): refactor code to trigger stack-analyses for manifest files
