+++ src/extension.ts
@@ -18,91 +18,7 @@ export function activate(context: vscode.ExtensionContext) {
 	let provider = new contentprovidermodule.TextDocumentContentProvider();  //new TextDocumentContentProvider();
   let registration = vscode.workspace.registerTextDocumentContentProvider('fabric8-analytics-widget', provider);
 
-	let disposable = vscode.commands.registerCommand(Commands.TRIGGER_STACK_ANALYSIS, () => {
-    if(vscode.workspace.hasOwnProperty('workspaceFolders') && vscode.workspace['workspaceFolders'].length>1){
-      vscode.window.showInformationMessage('Multi-root Workspaces are not supported currently');
-    } else if(vscode.window.activeTextEditor){
-      let editor = vscode.window.activeTextEditor;
-      let text = editor.document.getText();
-      let fileUri: string = editor.document.fileName;
-
-      vscode.window.withProgress({ location: vscode.ProgressLocation.Window, title: 'Generate application stack report'}, p => {
-        return new Promise((resolve, reject) => {       
-          if(fileUri.toLowerCase().indexOf('pom.xml')!== -1){
-            p.report({message: 'Generating effective pom ...' });
-            ProjectDataProvider.effectivef8Pom(editor.document.uri, (dataEpom) => {
-                if(dataEpom){
-                  p.report({message: 'Analyzing your stack ...' });
-                    provider.signalInit(previewUri,null);
-                      authextension.authorize_f8_analytics(context, (data) => {
-                        if(data){
-                          return vscode.commands.executeCommand('vscode.previewHtml', previewUri, vscode.ViewColumn.One, 'Application stack report').then((success) => {
-                            stackanalysismodule.get_stack_metadata(context, dataEpom, {manifest: text, origin: 'lsp'}, provider, Apiendpoint.OSIO_ACCESS_TOKEN, (data) => {
-                              if(data){
-                                p.report({message: 'Successfully generated stack report ...' });
-                                resolve();
-                                provider.signal(previewUri, data); 
-                              } else {
-                                provider.signal(previewUri,null);
-                                reject();
-                              }
-                            });
-                            provider.signalInit(previewUri,null);
-                          }, (reason) => {
-                                reject();
-                                vscode.window.showErrorMessage(reason);
-                          });
-                        } else {
-                            reject();
-                        }
-                    });
-              } else {
-                  p.report({message: 'Unable to generate effective pom ...' });
-                  reject();
-                }
-            });
-          } else {
-            ProjectDataProvider.effectivef8Package(editor.document.uri.fsPath, (ePkgPath) => {
-            if(ePkgPath){
-              p.report({message: 'Analyzing your stack ...' });
-              provider.signalInit(previewUri,null);
-              authextension.authorize_f8_analytics(context, (data) => {
-                if(data){
-                  return vscode.commands.executeCommand('vscode.previewHtml', previewUri, vscode.ViewColumn.One, 'Application stack report').then((success) => {
-                    stackanalysismodule.get_stack_metadata(context, ePkgPath, {manifest: text, origin: 'lsp'}, provider, Apiendpoint.OSIO_ACCESS_TOKEN, (data) => {
-                      if(data){
-                        p.report({message: 'Successfully generated stack report ...' });
-                        resolve();
-                        provider.signal(previewUri, data);
-                      } else {
-                        provider.signal(previewUri,null);
-                        reject();
-                      }
-                    });
-                    provider.signalInit(previewUri,null);
-                  }, (reason) => {
-                    reject();
-                    vscode.window.showErrorMessage(reason);
-                  });
-                } else {
-                    reject();
-                  }
-              });
-            }
-            else {
-              p.report({message: 'Unable to resolve dependencies in package.json ...' });
-              reject();
-            }
-            });
-
-          }
-
-        });
-      });
-    } else {
-      vscode.window.showInformationMessage('No manifest file is active in editor');
-    }
-	});
+	let disposable = vscode.commands.registerCommand(Commands.TRIGGER_STACK_ANALYSIS, () => stackanalysismodule.triggerStackAnalyses(context, provider, previewUri));
 
   let disposableFullStack = vscode.commands.registerCommand(Commands.TRIGGER_FULL_STACK_ANALYSIS, () => {
     if(vscode.workspace.hasOwnProperty('workspaceFolders') && vscode.workspace['workspaceFolders'].length>1){
